# backend/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Raptr Automation API",
    "description": "Automation + Manual Test management API",
    "version": "1.0.0"
  },
  "host": "localhost:6223",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/api/bugzilla/bugs/": {
      "get": {
        "description": "",
        "produces": [
          "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "type": "string"
          },
          {
            "name": "limit",
            "in": "query",
            "type": "string"
          },
          {
            "name": "sortBy",
            "in": "query",
            "type": "string"
          },
          {
            "name": "sortOrder",
            "in": "query",
            "type": "string"
          },
          {
            "name": "download",
            "in": "query",
            "type": "string"
          },
          {
            "name": "sendMail",
            "in": "query",
            "type": "string"
          },
          {
            "name": "filters",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/bugzilla/bugs/{bugId}/exists": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "bugId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/bugzilla/bugs/batch": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "ids",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/bugzilla/bugs/testcase": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "testcaseIds",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/bugzilla/bugs/testcaseByRunId": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "type": "string"
          },
          {
            "name": "limit",
            "in": "query",
            "type": "string"
          },
          {
            "name": "sortBy",
            "in": "query",
            "type": "string"
          },
          {
            "name": "sortOrder",
            "in": "query",
            "type": "string"
          },
          {
            "name": "download",
            "in": "query",
            "type": "string"
          },
          {
            "name": "sendMail",
            "in": "query",
            "type": "string"
          },
          {
            "name": "runId",
            "in": "query",
            "type": "string"
          },
          {
            "name": "filters",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/{bugId}/exists": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "bugId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/batch": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "ids",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/testcase": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "testcaseIds",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/testcaseByRunId": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "type": "string"
          },
          {
            "name": "limit",
            "in": "query",
            "type": "string"
          },
          {
            "name": "sortBy",
            "in": "query",
            "type": "string"
          },
          {
            "name": "sortOrder",
            "in": "query",
            "type": "string"
          },
          {
            "name": "download",
            "in": "query",
            "type": "string"
          },
          {
            "name": "sendMail",
            "in": "query",
            "type": "string"
          },
          {
            "name": "runId",
            "in": "query",
            "type": "string"
          },
          {
            "name": "filters",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  }
}